requirements.txt
asgiref
Django
djangorestframework
drf-nested-routers
pytz
sqlparse
typing_extensions
mysqlclient
python-dotenv
